seperated game logic from board
